--- src/main/java/kr/hhplus/be/server/common/lock/LockCallBack.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.common.lock;
-
-@FunctionalInterface
-public interface LockCallBack<T> {
-    T call() throws Throwable;
-}
