--- src/main/java/kr/hhplus/be/server/domain/payment/PaymentDataPlatformClient.java
@@ -1,6 +1,0 @@
-package kr.hhplus.be.server.domain.payment;
-
-public interface PaymentDataPlatformClient {
-
-    void sendPaymentData(Payment payment);
-}
