+++ src/main/java/kr/hhplus/be/server/infrastructure/lock/LockManager.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.common.lock;
+package kr.hhplus.be.server.infrastructure.lock;
 
 public interface LockManager extends AutoCloseable {
 }
